Modification of the format of the patterns (1 bit less for module position and 1 bit less for superstrip position) Maximum number of modules on the same ladder in a sector is now 16, the smallest superstrip size is 32 strips
It is now possible to use the PT of the stubs in the patterns. In that case, the third DC bit is kept for that usage (letting only 2 DC bits available for the superstrips)
The Stub PT DC bit is set to 0 if ALL tracks leading to this pattern had a stub's PT <= -1.5 on this layer. It is set to 1 if ALL tracks leading to this pattern had a stub's PT >= 1.5 on this layer. In all other cases the DC bit is set to X and the information is not used for the pattern recognition.

To generate a bank using the Stub's PT, add the --useStubPT parameter. Same parameter must be used while performing the pattern recognition, otherwise the stub's PT information will not be used.